requirements.txt
Django==5.0.2
djangorestframework==3.14.0
pandas==2.2.3
openai==1.77.0
gunicorn==21.0.1
psycopg2-binary==2.9.10
django-cors-headers==4.2.0
channels==4.2.2